BookMacster 1.21 (was 1.20.7), prior to Code Review.   After updating to Xcode 5.1 which wanted lots of stupid integer typecasts, finally fixed the #import Sparkle/SUUpdater.h phenomena (removed the path), added “Periodic” advanced trigger.

+++ SSYDocFileObserver.m
@@ -72,7 +72,7 @@ NSString* const SSYDocFileObserverErrorDomain = @"SSYDocFileObserverErrorDomain"
              The following was changed in BookMacster 1.20.2, to catch updates
              by both Dropbox and ChronoSync.
              */
-            NSInteger watchFlags = (SSYPathObserverChangeFlagsRename | SSYPathObserverChangeFlagsDelete) ;
+            uint32_t watchFlags = (SSYPathObserverChangeFlagsRename | SSYPathObserverChangeFlagsDelete) ;
 
 			ok = [[self pathObserver] addPath:path
 								   watchFlags:watchFlags
